<Feat> : 최상위 조직 생성시 apt 저장소 생성 및 패키지 등록 추가

+++ hamonize-center/src/main/java/com/mapper/IFileMapper.java
@@ -7,7 +7,7 @@ public interface IFileMapper {
     
     public List <FileVo> getFiles();
     public FileVo getFile(String type);
-    public FileVo getFileSeq(int seq);
+    public FileVo getFileSeq(Long seq);
     public FileVo getFilePublic();
     public FileVo getFilePrivate();
     public FileVo getFileConfig();
